`passwd` interface
This adds a passwd interface to allow the user to change their password without needing to have suid on the `passwd` binary. <- Not sure how susceptible this is to a non-root brute force.

I also removed the ability for users to authenticate themselves. `su` and `sudo` need to be root anyway to change the uid of the current process. Also, any random userspace binary can basically do a brute force attack on your machine without needing to be root if you're allowed to authenticate yourself.